update api architecture
- unified a helper function get_token_with_roles
- use get_token_with_roles in test_groups
- delayed imports in conftest and shared.helpers to circumvent circular
import error
